requirements.txt
pandas~=1.5.3
scikit-learn~=1.2.1
category_encoders~=2.6.0
sqlalchemy~=1.4.39
psycopg2~=2.9.9
python-dotenv~=1.0.0
pika~=1.3.2
flask~=2.2.2
joblib~=1.1.1